fixing code

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -6,13 +6,12 @@
  */
 size_t list_len(const list_t *h)
 {
-	int i;
+	size_t i;
 
 	while (h != NULL)
 	{
 		h = h->next;
 		i++;
 	}
-return (i);
-
+	return (i);
 }
